apps/api/pyproject.toml
[project]
name = "korb-api"
version = "0.2.0"
description = "FastAPI backend for Korb meal planning app"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.0",
    "email-validator>=2.1.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.115.0",
    "fastembed>=0.4.0",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.6.0",
    "posthog>=3.0.0",
    "pyjwt[crypto]>=2.8.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.12.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "async-factory-boy>=1.0.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
    "testcontainers[postgres]>=4.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
